# Vite: Add jsdoc-type-pratt-parser to `optimizeDeps`